fix(broker): issue due to backport

+++ broker/core/inc/com/centreon/broker/misc/perfdata.hh
@@ -89,7 +89,7 @@ class perfdata {
  *  @return Metric value.
  */
 // Inlined after profiling for performance.
-inline double perfdata::value() const noexcept {
+inline float perfdata::value() const noexcept {
   return _value;
 }
 }  // namespace misc
